+++ CHANGES.md
@@ -7,4 +7,4 @@
 - Deactivate Voids
 - Fixed missing Export table on Payouts
 - Option to hide Gross Amount in Payouts
-- 
+- Removed complex Acquirer Rules for Standard Merchants